@@ -4,11 +4,11 @@
   "private": true,
   "dependencies": {
     "orbit-controls-es6": "^2.0.1",
-    "react": "^16.11.0",
-    "react-dom": "^16.11.0",
-    "react-scripts": "3.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "^5.0.1",
     "serve": "^11.2.0",
-    "three": "^0.110.0",
+    "three": "^0.143.0",
     "three-mtl-loader": "^1.0.2",
     "three-obj-loader": "^1.1.3",
     "three-obj-loader-es6-module": "^1.0.1",
